Adds public scope to GraphQL queries
Adds the `PUBLIC` scope to several GraphQL queries and mutations to allow unauthenticated users to access them.

Also, this commit updates the graphql version in manifest.json and upgrades some dependencies in the node and react `package.json` files, and it also changes the import paths for types from `vtex.return-app` to `baranda.return-app-pmi`.